Implemented Input manager, BIG CHANGES read description
I also changed the player sprite filter to point mode to make it look
sharper, as well as changed the player collider to a poly collider. A
lot of problems/bugs came from many changes I believe most of them have
to do with the scaling that the tutorial uses being different from our
scene.

List of issues:
- When FaceDirection is active the player sprite grows larger I can’t
figure out why.
- There is a strange delay on horizontal movement that causes the
character to continue walking for a while after the key is lifted.
- When jumping the CollisionState script is supposed to restrict jumps
but is not working because I have not implemented a PlayerManager script
that the tutorial uses to handle animation/behavior states.
I did not want to change the scaling until I had Aaron go over this and
see if he can find what is wrong.