Changed back to utilizing CameraFollow.
CameraSeek and CameraFollow ended up becoming the same. I just added the
ChangeYBounds() method on CameraFollow.

To use:
A Camera Zone Object is placed on the game environment. The
CameraBoundsSwitching Script is then placed on it, where it would call
the ChangeYBounds() to the objects position and scale.

Could use more work to be smoother and more accurate, but for now it's
functional.

-Also, changed the camera to work work on update instead of fixedupdate
(may give us issues later on, but for now it's much smoother.)