Attempted to fix Camera.
Created CameraScript and CameraTransitionScript.

CameraScript: Attach to camera to have it respond to player and camera
transition areas.

CameraTransitionScript: On player enter, call CameraScripts
SetClamps(Transform newArea) method. This changes the camera limits to
the transform of the camera transition game object.
-Note: Camera transition areas must be overlapped with one another,
otherwise, it becomes jittery.

Still working on smoothening the transition.