Camera Transition is better without overlap.
Still trying to figure out if the camera should move into the next zones
or if snapping to the next zone is just fine.

Also found a bug with PathFinding. Due to a misalignment of times of
moveDuration and scanDuration, the enemy can still move even though that
direction shouldn't be valid anymore. (Happens about 10% of the time.)
For now just leaving it alone. Will either have to play with the move
and scan duration times, or have move be more limited by scan.